created calculator dir; moved relevant component files; created Display component
