+++ src/calculator/Calculator.css
@@ -11,22 +11,6 @@ p {
     padding: 10px;
 }
 
-.display {
-    border: #3c4043 1px solid;
-    border-radius: 2px;
-    color: #e8eaed;
-    cursor: pointer;
-    margin: 4px;
-    text-align: right;
-    width: calc(100% - 8px);
-}
-
-.display span {
-    line-height: 44px;
-    font-size: 22px;
-    padding: 10px;
-}
-
 .buttons button{
     background-color: #3c4043;
     border: 1px solid #3c4043;
@@ -40,7 +24,7 @@ p {
     width: 85px;
 }
 
-.buttons button:hover, .display:hover{
+.buttons button:hover{
     background: #424548;
     border-color: #424548;
 }
